Update WG2 API to reflect established Weight Semantics
The docstrings have been updated to reflect the correct node and edge
weight calculation semantics.

Requested getters have also been added
